+++ src/lr-wpan/model/lr-wpan-mac.cc
@@ -24,6 +24,7 @@
  */
 #include "lr-wpan-mac.h"
 
+#include "lr-wpan-constants.h"
 #include "lr-wpan-csmaca.h"
 #include "lr-wpan-mac-header.h"
 #include "lr-wpan-mac-pl-headers.h"
@@ -45,51 +46,6 @@ namespace ns3
 
 NS_LOG_COMPONENT_DEFINE("LrWpanMac");
 NS_OBJECT_ENSURE_REGISTERED(LrWpanMac);
-
-/**
- * \defgroup LrWpanMacSublayerConstants MAC sublayer constants
- * \ingroup lr-wpan
- * @{
- */
-
-/**
- * The minimum number of octets added by the MAC sublayer to the PSDU.
- * See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aMinMPDUOverhead = 9;
-
-/**
- * Length of a superframe slot in symbols. Defaults to 60 symbols in each superframe slot.
- * See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aBaseSlotDuration = 60;
-
-/**
- * Number of a superframe slots per superframe. Defaults to 16.
- * See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aNumSuperframeSlots = 16;
-
-/**
- * Length of a superframe in symbols. Defaults to aBaseSlotDuration * aNumSuperframeSlots in
- * symbols. See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aBaseSuperframeDuration = aBaseSlotDuration * aNumSuperframeSlots;
-
-/**
- * The number of consecutive lost beacons that will cause the MAC sublayer of a receiving device to
- * declare a loss of synchronization.
- * See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aMaxLostBeacons = 4;
-
-/**
- * The maximum size of an MPDU, in octets, that can be followed by a Short InterFrame Spacing (SIFS)
- * period. See IEEE 802.15.4-2011, section 6.4.1, Table 51.
- */
-constexpr uint32_t aMaxSIFSFrameSize = 18;
-
-/**@}*/
 
 TypeId
 LrWpanMac::GetTypeId()
@@ -247,7 +203,7 @@ LrWpanMac::LrWpanMac()
     m_channelScanIndex = 0;
     m_maxEnergyLevel = 0;
 
-    m_macResponseWaitTime = aBaseSuperframeDuration * 32;
+    m_macResponseWaitTime = lrwpan::aBaseSuperframeDuration * 32;
     m_assocRespCmdWaitTime = 960;
 
     m_maxTxQueueSize = m_txQueue.max_size();
@@ -393,7 +349,7 @@ LrWpanMac::McpsDataRequest(McpsDataRequestParams params, Ptr<Packet> p)
     LrWpanMacHeader macHdr(LrWpanMacHeader::LRWPAN_MAC_DATA, m_macDsn.GetValue());
     m_macDsn++;
 
-    if (p->GetSize() > LrWpanPhy::aMaxPhyPacketSize - aMinMPDUOverhead)
+    if (p->GetSize() > lrwpan::aMaxPhyPacketSize - lrwpan::aMinMPDUOverhead)
     {
         // Note, this is just testing maximum theoretical frame size per the spec
         // The frame could still be too large once headers are put on
@@ -840,7 +796,8 @@ LrWpanMac::MlmeSyncRequest(MlmeSyncRequestParams params)
     {
         m_numLostBeacons = 0;
         // search for a beacon for a time = incomingSuperframe symbols + 960 symbols
-        searchSymbols = ((uint64_t)1 << m_incomingBeaconOrder) + 1 * aBaseSuperframeDuration;
+        searchSymbols =
+            ((uint64_t)1 << m_incomingBeaconOrder) + 1 * lrwpan::aBaseSuperframeDuration;
         searchBeaconTime = Seconds((double)searchSymbols / symbolRate);
         m_beaconTrackingOn = true;
         m_trackingEvent =
@@ -1185,9 +1142,9 @@ LrWpanMac::EndStartRequest()
             m_fnlCapSlot = 15;
 
             m_beaconInterval =
-                (static_cast<uint32_t>(1 << m_macBeaconOrder)) * aBaseSuperframeDuration;
-            m_superframeDuration =
-                (static_cast<uint32_t>(1 << m_macSuperframeOrder)) * aBaseSuperframeDuration;
+                (static_cast<uint32_t>(1 << m_macBeaconOrder)) * lrwpan::aBaseSuperframeDuration;
+            m_superframeDuration = (static_cast<uint32_t>(1 << m_macSuperframeOrder)) *
+                                   lrwpan::aBaseSuperframeDuration;
 
             // TODO: change the beacon sending according to the startTime parameter (if not PAN
             // coordinator)
@@ -1460,7 +1417,7 @@ LrWpanMac::BeaconSearchTimeout()
 {
     uint64_t symbolRate = (uint64_t)m_phy->GetDataOrSymbolRate(false); // symbols per second
 
-    if (m_numLostBeacons > aMaxLostBeacons)
+    if (m_numLostBeacons > lrwpan::aMaxLostBeacons)
     {
         MlmeSyncLossIndicationParams syncLossParams;
         // syncLossParams.m_logCh =
@@ -1478,7 +1435,8 @@ LrWpanMac::BeaconSearchTimeout()
         // Search for one more beacon
         uint64_t searchSymbols;
         Time searchBeaconTime;
-        searchSymbols = ((uint64_t)1 << m_incomingBeaconOrder) + 1 * aBaseSuperframeDuration;
+        searchSymbols =
+            ((uint64_t)1 << m_incomingBeaconOrder) + 1 * lrwpan::aBaseSuperframeDuration;
         searchBeaconTime = Seconds((double)searchSymbols / symbolRate);
         m_trackingEvent =
             Simulator::Schedule(searchBeaconTime, &LrWpanMac::BeaconSearchTimeout, this);
@@ -1830,7 +1788,7 @@ LrWpanMac::PdDataIndication(uint32_t psduLength, Ptr<Packet> p, uint8_t lqi)
                 if (acceptFrame && (m_csmaCa->IsSlottedCsmaCa() && m_capEvent.IsRunning()))
                 {
                     Time timeLeftInCap = Simulator::GetDelayLeft(m_capEvent);
-                    uint64_t ackSymbols = m_phy->aTurnaroundTime + m_phy->GetPhySHRDuration() +
+                    uint64_t ackSymbols = lrwpan::aTurnaroundTime + m_phy->GetPhySHRDuration() +
                                           ceil(6 * m_phy->GetPhySymbolsPerOctet());
                     Time ackTime = Seconds((double)ackSymbols / symbolRate);
 
@@ -1980,9 +1938,9 @@ LrWpanMac::PdDataIndication(uint32_t psduLength, Ptr<Packet> p, uint8_t lqi)
 
                         m_incomingBeaconInterval =
                             (static_cast<uint32_t>(1 << m_incomingBeaconOrder)) *
-                            aBaseSuperframeDuration;
+                            lrwpan::aBaseSuperframeDuration;
                         m_incomingSuperframeDuration =
-                            aBaseSuperframeDuration *
+                            lrwpan::aBaseSuperframeDuration *
                             (static_cast<uint32_t>(1 << m_incomingSuperframeOrder));
 
                         if (incomingSuperframe.IsBattLifeExt())
@@ -2084,7 +2042,7 @@ LrWpanMac::PdDataIndication(uint32_t psduLength, Ptr<Packet> p, uint8_t lqi)
 
                                 searchSymbols =
                                     (static_cast<uint64_t>(1 << m_incomingBeaconOrder)) +
-                                    1 * aBaseSuperframeDuration;
+                                    1 * lrwpan::aBaseSuperframeDuration;
                                 searchBeaconTime =
                                     Seconds(static_cast<double>(searchSymbols / symbolRate));
                                 m_trackingEvent =
@@ -2552,12 +2510,12 @@ LrWpanMac::EnqueueInd(Ptr<Packet> p)
     if (m_macBeaconOrder == 15)
     {
         // Non-beacon enabled mode
-        unit = aBaseSuperframeDuration * m_macTransactionPersistenceTime;
+        unit = lrwpan::aBaseSuperframeDuration * m_macTransactionPersistenceTime;
     }
     else
     {
         // Beacon-enabled mode
-        unit = ((static_cast<uint32_t>(1) << m_macBeaconOrder) * aBaseSuperframeDuration) *
+        unit = ((static_cast<uint32_t>(1) << m_macBeaconOrder) * lrwpan::aBaseSuperframeDuration) *
                m_macTransactionPersistenceTime;
     }
 
@@ -3104,7 +3062,7 @@ LrWpanMac::PlmeSetAttributeConfirm(LrWpanPhyEnumeration status, LrWpanPibAttribu
         if (status == LrWpanPhyEnumeration::IEEE_802_15_4_PHY_SUCCESS)
         {
             uint64_t symbolRate = static_cast<uint64_t>(m_phy->GetDataOrSymbolRate(false));
-            uint64_t scanDuration = aBaseSuperframeDuration *
+            uint64_t scanDuration = lrwpan::aBaseSuperframeDuration *
                                     ((static_cast<uint32_t>(1 << m_scanParams.m_scanDuration)) + 1);
             Time nextScanTime = Seconds(static_cast<double>(scanDuration / symbolRate));
 
@@ -3477,7 +3435,7 @@ LrWpanMac::ChangeMacState(LrWpanMacState newState)
 uint64_t
 LrWpanMac::GetMacAckWaitDuration() const
 {
-    return m_csmaCa->GetUnitBackoffPeriod() + m_phy->aTurnaroundTime + m_phy->GetPhySHRDuration() +
+    return m_csmaCa->GetUnitBackoffPeriod() + lrwpan::aTurnaroundTime + m_phy->GetPhySHRDuration() +
            ceil(6 * m_phy->GetPhySymbolsPerOctet());
 }
 
@@ -3528,7 +3486,7 @@ LrWpanMac::GetIfsSize()
 {
     NS_ASSERT(m_txPkt);
 
-    if (m_txPkt->GetSize() <= aMaxSIFSFrameSize)
+    if (m_txPkt->GetSize() <= lrwpan::aMaxSIFSFrameSize)
     {
         return m_macSIFSPeriod;
     }